Added CaseList

+++ app/src/main/java/com/example/csgocaseswatcherapp/domain/Case.kt
@@ -4,8 +4,8 @@ data class Case(
     val name: String,
     val releaseDate: String,
     val dropStatus: String,
-    val lowestPrice: Float,
+    val lowestPrice: Double,
     val volume: Int,
-    val medianPrice: Float,
+    val medianPrice: Double,
     val imageUrl: String
 )
